Made changes to advmdata.nortek module
